requirements.txt
# 基础运行环境
streamlit>=1.28.0
pandas>=1.5.0
numpy>=1.21.0
openai>=1.0.0
plotly>=5.15.0

# Excel处理
openpyxl>=3.1.0
xlrd>=2.0.0 

# 文档分析依赖
markitdown[all]>=0.1.0  # 包含PDF、DOCX等所有格式支持
python-docx>=0.8.11     # DOCX文档处理
PyPDF2>=3.0.0           # PDF文档处理
pymupdf>=1.23.0         # 更强大的PDF处理（fitz）
Pillow>=9.0.0           # 图像处理

# 文档分析的可选依赖（提升处理能力）
pdfplumber>=0.9.0       # PDF表格和布局分析
docx2txt>=0.8           # DOCX纯文本提取
textract>=1.6.5         # 多格式文档提取（可选） 